setup users route and adjusted endpoints
