+++ data/middleware/middleware.js
@@ -11,7 +11,7 @@ function restricted(req, res, next) {
 			}
 		});
 	} else {
-		res.status(401).json({ message: 'You shall not pass' });
+		res.status(500).json({ message: 'You shall not pass' });
 	}
 }
